backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.1
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.4.2
fastapi-cors==0.4.0
